+++ smalisca/modules/module_graph.py
@@ -34,7 +34,7 @@
 
 import graphviz as gv
 
-from smalisca.core.smalisca_config import GraphConfig
+from smalisca.core.smalisca_config import smalisca_conf
 from smalisca.core.smalisca_logging import log
 
 
@@ -205,6 +205,7 @@ class ClassGraph(GraphBase):
         self.classes = {}
         self.subgraphs = {}
         self.edges = []
+        self.graphviz_opts = smalisca_conf.options['graphviz']
 
     def add_class(self, class_obj):
         """Add new class object to graph
@@ -260,8 +261,8 @@ class ClassGraph(GraphBase):
         node_attr = {'label': class_label}
 
         # Add methode node attributes
-        for k in GraphConfig.ClassGraphConfig.class_nodes['nodes'].keys():
-            node_attr[k] = GraphConfig.ClassGraphConfig.class_nodes['nodes'][k]
+        for k in self.graphviz_opts['classes']['class_nodes']['nodes'].keys():
+            node_attr[k] = self.graphviz_opts['classes']['class_nodes']['nodes'][k]
 
         add_nodes(package_graph, [(class_node, node_attr)])
 
@@ -285,15 +286,14 @@ class ClassGraph(GraphBase):
 
         # First make subgraphs
         for k in self.subgraphs.keys():
-
             # Add styles
-            apply_styles(self.subgraphs[k], GraphConfig.ClassGraphConfig.cluster_styles)
+            apply_styles(self.subgraphs[k], self.graphviz_opts['classes']['cluster_styles'])
 
             # Make subgraph
             self.G.subgraph(self.subgraphs[k])
 
         # Add global graph styles
-        apply_styles(self.G, GraphConfig.ClassGraphConfig.graph_styles)
+        apply_styles(self.G, self.graphviz_opts['classes']['graph_styles'])
 
 
 class CallGraph(GraphBase):
@@ -314,6 +314,7 @@ class CallGraph(GraphBase):
         self.subgraphs = {}
         self.edges = []
         self.classes = {}
+        self.graphviz_opts = smalisca_conf.options['graphviz']
 
     def add_class_subgraph(self, class_node):
         """Adds a new subgraph depending on the node
@@ -383,8 +384,8 @@ class CallGraph(GraphBase):
             method_node_attr = {'label': "[M] %s\l" % call_obj.from_method}
 
             # Add additional methode node attributes
-            for k in GraphConfig.CallsGraphConfig.method_nodes['nodes'].keys():
-                method_node_attr[k] = GraphConfig.CallsGraphConfig.method_nodes['nodes'][k]
+            for k in self.graphviz_opts['calls']['method_nodes']['nodes'].keys():
+                method_node_attr[k] = self.graphviz_opts['calls']['method_nodes']['nodes'][k]
 
             # Add node
             add_nodes(from_subgraph, [(
@@ -398,8 +399,8 @@ class CallGraph(GraphBase):
             method_node_attr = {'label': "[M] %s\l" % call_obj.dst_method}
 
             # Add methode node attributes
-            for k in GraphConfig.CallsGraphConfig.method_nodes['nodes'].keys():
-                method_node_attr[k] = GraphConfig.CallsGraphConfig.method_nodes['nodes'][k]
+            for k in self.graphviz_opts['calls']['method_nodes']['nodes'].keys():
+                method_node_attr[k] = self.graphviz_opts['calls']['method_nodes']['nodes'][k]
 
             # Add node
             add_nodes(to_subgraph, [(
@@ -412,8 +413,8 @@ class CallGraph(GraphBase):
             edge_attr = {}
 
             # Add edge attributes
-            for k in GraphConfig.CallsGraphConfig.method_edges['edges'].keys():
-                edge_attr[k] = GraphConfig.CallsGraphConfig.method_edges['edges'][k]
+            for k in self.graphviz_opts['calls']['method_edges']['edges'].keys():
+                edge_attr[k] = self.graphviz_opts['calls']['method_edges']['edges'][k]
 
             # Add edge
             add_edges(self.G, [
@@ -432,10 +433,10 @@ class CallGraph(GraphBase):
         for k in self.classes.keys():
 
             # Apply styles
-            apply_styles(self.classes[k]['graph'], GraphConfig.CallsGraphConfig.cluster_styles)
+            apply_styles(self.classes[k]['graph'], self.graphviz_opts['calls']['cluster_styles'])
 
             # Make subgraph
             self.G.subgraph(self.classes[k]['graph'])
 
         # Apply general graph styles
-        apply_styles(self.G, GraphConfig.CallsGraphConfig.graph_styles)
+        apply_styles(self.G, self.graphviz_opts['calls']['graph_styles'])